Fix: Resolve InvalidCastException for WrapPanel ItemWidth/Height
This commit addresses two main points:

1.  **Fix InvalidCastException for WrapPanel Item Sizing:**
    - I changed the types of `DesignWidth` and `DesignHeight` static fields in `UserControls/ChapterListElement.xaml.cs` from `int` to `double`.
    - This resolves a `System.InvalidCastException` that occurred when using these integer values directly with `WrapPanel.ItemWidth` and `WrapPanel.ItemHeight` properties (which expect doubles) via `x:Static` binding in XAML.

2.  **Add Diagnostic Logging for Chapter Loading:**
    - I added `System.Diagnostics.Debug.WriteLine` statements to the `LoadChapterListAsync` method in `MainPage/MainWindow.xaml.cs`.
    - This logging was introduced to help diagnose issues with items not appearing after the refactor to `ItemsControl` (which was subsequently traced to the type cast issue and XAML configuration).

+++ UserControls/ChapterListElement.xaml.cs
@@ -19,8 +19,8 @@ namespace Reader.UserControls
         private DirectoryData _directory { get; set; }
         private List<string>? _imagePaths = null; // Made nullable and initialized to null
         public static readonly int ImageHeight = 250;
-        public static readonly int DesignHeight = 350;
-        public static readonly int DesignWidth = 199;
+        public static readonly double DesignHeight = 350.0;
+        public static readonly double DesignWidth = 199.0;
 
 
         public ChapterListElement(DirectoryInfo directoryInfo)
